projeto

+++ app/src/main/java/com/example/pulsesocial/feature/components/PostCard.kt
@@ -1,16 +1,17 @@
 package com.example.pulsesocial.feature.components
 
-import android.content.Context
 import android.os.Build
 import androidx.annotation.RequiresApi
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.ChatBubbleOutline
@@ -29,11 +30,10 @@ import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.example.pulsesocial.data.datastore.UserPreferencesRepository
-import com.example.pulsesocial.data.session.SessionManager
+import coil3.compose.AsyncImage
 import com.example.pulsesocial.domain.response.PostResponse
 import com.example.pulsesocial.domain.response.UserSummary
 import com.example.pulsesocial.ui.theme.PulseSocialTheme
@@ -41,6 +41,7 @@ import com.example.pulsesocial.ui.theme.PulseSocialTheme
 @Composable
 fun PostCard(
     post: PostResponse,
+    currentUserId: Long,
     onLikeClick: () -> Unit,
     onDeletePostClick: () -> Unit,
 ) {
@@ -67,19 +68,35 @@ fun PostCard(
                     .fillMaxWidth()
             ) {
 
-                Surface(
-                    modifier = Modifier
-                        .padding(end = 16.dp)
-                        .size(50.dp),
-                    color = MaterialTheme.colorScheme.background,
-                    shape = RoundedCornerShape(60)
+                if (post.author.imageUrl.isEmpty()) {
+                    Surface(
+                        modifier = Modifier
+                            .padding(end = 16.dp)
+                            .size(50.dp),
+                        color = MaterialTheme.colorScheme.background,
+                        shape = RoundedCornerShape(60)
 
-                ) {
-                    Icon(
-                        imageVector = Icons.Default.Person,
-                        contentDescription = "Avatar",
-                        tint = MaterialTheme.colorScheme.primary
-                    )
+                    ) {
+                        Icon(
+                            imageVector = Icons.Default.Person,
+                            contentDescription = "Avatar",
+                            tint = MaterialTheme.colorScheme.primary
+                        )
+                    }
+                } else {
+                    Surface(
+                        modifier = Modifier
+                            .padding(end = 16.dp)
+                            .size(50.dp),
+                        shape = CircleShape
+                    ) {
+                        AsyncImage(
+                            model = post.author.imageUrl,
+                            contentDescription = "Foto de perfil",
+                            modifier = Modifier.fillMaxSize(),
+                            contentScale = ContentScale.Crop
+                        )
+                    }
                 }
 
                 Text(
@@ -88,20 +105,39 @@ fun PostCard(
 
                 Spacer(Modifier.weight(1f))
 
-                IconButton(onClick = { onDeletePostClick()}) {
-                    Icon(
-                        imageVector = Icons.Default.Delete,
-                        contentDescription = "Delete",
-                        tint = MaterialTheme.colorScheme.primary
-                    )
+                if (post.userId == currentUserId) {
+
+                    IconButton(onClick = { onDeletePostClick() }) {
+                        Icon(
+                            imageVector = Icons.Default.Delete,
+                            contentDescription = "Delete",
+                            tint = MaterialTheme.colorScheme.primary
+                        )
+                    }
                 }
 
             }
 
-            Text(
-                modifier = Modifier.padding(top = 16.dp),
-                text = post.content
-            )
+            if (post.imageUrl.isEmpty()) {
+                Text(
+                    modifier = Modifier.padding(top = 16.dp),
+                    text = post.content
+                )
+            } else {
+
+                AsyncImage(
+                    model = post.imageUrl,
+                    contentDescription = "Foto no Post",
+                    modifier = Modifier
+                        .padding(top = 16.dp)
+                        .fillMaxWidth()
+                )
+
+                Text(
+                    modifier = Modifier.padding(top = 16.dp),
+                    text = post.content
+                )
+            }
 
             Spacer(modifier = Modifier.height(8.dp))
 
@@ -160,7 +196,7 @@ fun PostCardPreview() {
         author = UserSummary(
             1,
             "Gabriel",
-            "img"
+            "https://doogspet.com/wp-content/uploads/freepik_edit-11-1.jpg"
         )
     )
 
@@ -169,6 +205,7 @@ fun PostCardPreview() {
     ) {
         PostCard(
             post = mockPost,
+            currentUserId = 0,
             onLikeClick = {},
             onDeletePostClick = {}
         )
